fix: import VolumeTransferred section Framer Motion's components on client-side
